+++ src/index.ts
@@ -6,13 +6,18 @@ export {
   Box,
   Center,
   Circle,
+  ColorModeScript,
+  cookieStorageManager,
   Flex,
   Grid,
   GridItem,
   HStack,
+  localStorageManager,
   Spacer,
   Square,
   Stack,
+  useColorMode,
+  useColorModeValue,
   VStack,
 } from "@chakra-ui/react";
 export {