+++ src/utils/componentCategories.ts
@@ -130,6 +130,7 @@ const categories = {
       "Breakpoints",
       "Buttons",
       "Colors",
+      "Color Mode",
       "Design Tokens",
       "Forms",
       "Iconography",